% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\name{markerNodes}
\alias{markerNodes}
\alias{find.marker.nodes}
\alias{markerNodes,SCFind,character-method}
\title{Performs query optimization and return the best candidate gene sets}
\usage{
markerNodes(object, node.list, datasets, log.message = 0)

find.marker.nodes(object, node.list, datasets, log.message = 0)

\S4method{markerNodes}{SCFind,character}(object, node.list, datasets, log.message = 0)
}
\arguments{
\item{object}{SCFind object}

\item{node.list}{A list of splicing nodes existing in the database}

\item{datasets}{the datasets of the objects to be considered}

\item{log.message}{whether to print a verbose message}
}
\value{
hierarchical list of queries and their respective scores
}
\description{
This function can be used with quite long gene lists
that otherwise would have no cell hits in the database
}
